+++ test/Hero.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel;
+
 namespace FluentlyHttpClient.Test
 {
 	/// <summary>
@@ -8,5 +10,17 @@ namespace FluentlyHttpClient.Test
 		public string Key { get; set; }
 		public string Name { get; set; }
 		public string Title { get; set; }
+	}
+
+	public enum HeroRole
+	{
+		[Description("assassin")]
+		Assassin,
+
+		[Description("fighter")]
+		Fighter,
+
+		[Description("warrior")]
+		Warrior,
 	}
 }